Move signature utilities to core
