+++ ethereum/core/src/main/java/tech/pegasys/artemis/core/signatures/MessageSignerService.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.datastructures.validator;
+package tech.pegasys.artemis.core.signatures;
 
 import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.artemis.bls.BLSSignature;
