+++ ethereum/statetransition/src/testFixtures/java/tech/pegasys/artemis/statetransition/BeaconChainUtil.java
@@ -27,12 +27,12 @@ import tech.pegasys.artemis.core.AttestationGenerator;
 import tech.pegasys.artemis.core.BlockProposalTestUtil;
 import tech.pegasys.artemis.core.StateTransition;
 import tech.pegasys.artemis.core.results.BlockImportResult;
+import tech.pegasys.artemis.core.signatures.MessageSignerService;
+import tech.pegasys.artemis.core.signatures.TestMessageSignerService;
 import tech.pegasys.artemis.datastructures.blocks.BeaconBlock;
 import tech.pegasys.artemis.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.artemis.datastructures.operations.Attestation;
 import tech.pegasys.artemis.datastructures.state.BeaconState;
-import tech.pegasys.artemis.datastructures.util.validator.TestMessageSignerService;
-import tech.pegasys.artemis.datastructures.validator.MessageSignerService;
 import tech.pegasys.artemis.protoarray.StubForkChoiceStrategy;
 import tech.pegasys.artemis.ssz.SSZTypes.SSZList;
 import tech.pegasys.artemis.statetransition.forkchoice.ForkChoiceUtil;
