+++ validator/client/src/main/java/tech/pegasys/artemis/validator/client/loader/ValidatorLoader.java
@@ -28,8 +28,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import tech.pegasys.artemis.bls.BLSKeyPair;
 import tech.pegasys.artemis.bls.BLSPublicKey;
-import tech.pegasys.artemis.core.Signer;
-import tech.pegasys.artemis.datastructures.validator.LocalMessageSignerService;
+import tech.pegasys.artemis.core.signatures.LocalMessageSignerService;
+import tech.pegasys.artemis.core.signatures.Signer;
 import tech.pegasys.artemis.util.config.ArtemisConfiguration;
 import tech.pegasys.artemis.validator.client.Validator;
 import tech.pegasys.artemis.validator.client.signer.ExternalMessageSignerService;
