+++ validator/client/src/test/java/tech/pegasys/artemis/validator/client/loader/ValidatorLoaderTest.java
@@ -24,7 +24,7 @@ import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import tech.pegasys.artemis.bls.BLSPublicKey;
-import tech.pegasys.artemis.datastructures.validator.LocalMessageSignerService;
+import tech.pegasys.artemis.core.signatures.LocalMessageSignerService;
 import tech.pegasys.artemis.util.config.ArtemisConfiguration;
 import tech.pegasys.artemis.validator.client.Validator;
 import tech.pegasys.artemis.validator.client.signer.ExternalMessageSignerService;
